+++ R/urbioconnect-package.R
@@ -8,13 +8,14 @@ NULL
 globalVariables(
   c(
     "area",
-    "distance",
+    "interpatch_distance",
     "effective_mesh_ha",
     "geometry",
     "patch_area_mean",
     "patch_area_total_ha",
     "patch_id",
     "prob_connectedness",
+    "scenario",
     "species",
     "value"
   )